test(qs): isolate Unicode boundary regressions
